Use typedstream-go for decoding attributedBody on Linux
