+++ chatdb/message.go
@@ -4,12 +4,10 @@
 package chatdb
 
 import (
-	"bytes"
 	"database/sql"
 	"fmt"
 	"log/slog"
 	"regexp"
-	"strings"
 
 	"github.com/pkg/errors"
 )
@@ -127,17 +125,4 @@ func (d *chatDB) GetMessage(messageID int, handleMap map[int]string) (string, bo
 		slog.Warn("no valid text or attributedBody for message", "messageID", messageID)
 	}
 	return fmt.Sprintf("[%s] %s: %s\n", date, handle, msg), valid, nil
-}
-
-func (d *chatDB) decodeTypedStream(s string) (string, error) {
-	cmd := d.execCommand("typedstream-decode")
-	cmd.Stdin = bytes.NewReader([]byte(s))
-	decodedBodyBytes, err := cmd.Output()
-	if err != nil {
-		return "", errors.Wrap(err, "decode attributedBody - POSSIBLE FIX: Add typedstream-decode to your system path (installed with bagoup)")
-	}
-	decodedBody := string(decodedBodyBytes)
-	decodedBody = _TypedStreamAttributeRE.ReplaceAllString(decodedBody, "")
-	decodedBody = _TypedStreamMultilineAttributeRE.ReplaceAllString(decodedBody, "")
-	return strings.TrimSuffix(decodedBody, "\n"), nil
 }
